fix: manage errorof users

+++ users/src/users/users.service.ts
@@ -4,6 +4,8 @@ import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
+import { RpcException } from '@nestjs/microservices';
+import { status } from '@grpc/grpc-js';
 
 @Injectable()
 export class UsersService {
@@ -24,19 +26,25 @@ export class UsersService {
   async findOne(id: number): Promise<User> {
     const user = await this.userRepository.findOneBy({ id });
     if (!user) {
-      throw new NotFoundException(`User #${id} not found`);
+      throw new RpcException({ code: status.NOT_FOUND, message: `User #${id} not found`});
     }
     return user;
   }
 
   async update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
-    await this.findOne(id);
+    const user = await this.findOne(id);
+    if (!user) {
+      throw new RpcException({ code: status.NOT_FOUND, message: `User #${id} not found`});
+    }
     await this.userRepository.update(id, updateUserDto);
     return this.findOne(id);
   }
 
   async remove(id: number): Promise<void> {
     const user = await this.findOne(id);
+    if (!user) {
+      throw new RpcException({ code: status.NOT_FOUND, message: `User #${id} not found`});
+    }
     await this.userRepository.remove(user);
   }
 }
